requirements.txt
accelerate==1.3.0
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2024.12.0
h11==0.14.0
huggingface-hub==0.28.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
onnx==1.17.0
onnxruntime==1.20.1
packaging==24.2
pillow==11.1.0
pkgconfig==1.5.5
protobuf==5.29.3
psutil==6.1.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
pyvips==2.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
